fix: fix ConsumeClaim context and refactor duration utils
