--- cmds/duration.go
@@ -1,8 +1,0 @@
-package cmds
-
-import "time"
-
-func parseDuration(s string) time.Duration {
-	d, _ := time.ParseDuration(s)
-	return d
-}
