+++ utils/duration.go
@@ -1,8 +1,8 @@
-package flow
+package utils
 
 import "time"
 
-func parseDuration(s string) time.Duration {
+func ParseDuration(s string) time.Duration {
 	d, _ := time.ParseDuration(s)
 	return d
 }
